+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Usage:
 * Fix #659: Update Fabric8 Kubernetes Client to v5.3.0
 * Fix #602: Add documentation regarding automatic Route generation
 * Fix #630: Document usage for `jkube.build.switchToDeployment` flag
+* Fix #647: Resource configuration can now add annotations and labels to ServiceAccount
 
 ### 1.2.0 (2021-03-31)
 * Fix #529: `.maven-dockerignore`, `.maven-dockerexclude`, `.maven-dockerinclude` are no longer supported